+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [Unreleased]
 ### Changed
 - `JSONDeserializer`: fixed obscure bugs in deserialization of classes implementing `Map` interface
+- `JSONDeserializer`: allow use of constructor with additional default parameters
 
 ## [4.2] - 2023-01-08
 ### Changed